Make the reset password handler set a cookie and redirect to a URL that does not expose sensitive information.
The reset-password link contains a potentially-sensitive token.  Although
the possibility for abuse is small since the token is invalidated after
it has been used, we don't want to leak the token in any Referrer: headers.

By storing the sensitive information in a cookie and immediately redirecting,
we should never see the token in a Referrer header.